requirements.txt
requests~=2.31.0
selenium~=4.11.2
lxml~=5.1.0
python-telegram-bot~=20.3
python-dotenv~=0.21.1
babel~=2.14.0
